commit-34

+++ app/Http/Controllers/Admin/ExcelMediaController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\ExcelMedia\StoreExcelMediaRequest;
 use App\Http\Requests\Admin\ExcelMedia\UpdateExcelMediaRequest;
 use App\Repositories\ExcelMediaRepository;
+use App\Repositories\ProductRepository;
 use App\Services\Media\MediaFileService;
 use Exception;
 use Illuminate\Support\Facades\DB;
@@ -13,10 +14,13 @@ use Illuminate\Support\Facades\DB;
 class ExcelMediaController extends Controller
 {
     private $excelMediaRepository;
+    private $productRepository;
 
-    public function __construct(ExcelMediaRepository $excelMediaRepository)
+    public function __construct(ExcelMediaRepository $excelMediaRepository,
+                                ProductRepository    $productRepository)
     {
         $this->excelMediaRepository = $excelMediaRepository;
+        $this->productRepository = $productRepository;
     }
 
     public function index()
@@ -60,13 +64,18 @@ class ExcelMediaController extends Controller
             DB::transaction(function () use ($request, $id) {
                 $excel_media = $this->excelMediaRepository->findById($id);
 
+                $checkImageIdForOthers = $this->productRepository->checkImageIdForOthers($excel_media['media_id']);
+
                 if ($request->hasFile('media')) {
                     $this->excelMediaRepository->update($request, null, $id);
                     $media_id = MediaFileService::publicUpload($request->file('media'),
                         'excel_media', null)->id;
                     $this->excelMediaRepository->addMedia($media_id, $excel_media->id);
-                    if ($excel_media->media) {
-                        $excel_media->media->delete();
+
+                    if ($checkImageIdForOthers <= 1) {
+                        if ($excel_media->media) {
+                            $excel_media->media->delete();
+                        }
                     }
                 } else {
                     $this->excelMediaRepository->update($request, $excel_media->media_id, $id);
@@ -87,14 +96,20 @@ class ExcelMediaController extends Controller
             DB::transaction(function () use ($id) {
                 $excel_media = $this->excelMediaRepository->findById($id);
 
-                if ($excel_media->media) {
-                    $excel_media->media->delete();
-                }
+                $checkImageIdForOthers = $this->productRepository->checkImageIdForOthers($excel_media['media_id']);
 
-                $excel_media->delete();
+                if ($checkImageIdForOthers <= 1) {
+                    if ($excel_media->media) {
+                        $excel_media->media->delete();
+                    }
+                    $excel_media->delete();
+
+                    newFeedback();
+                } else {
+                    newFeedback('پیام', 'این مدیا در محصولات اکسلی استفاده شده است', 'warning');
+                }
             });
             DB::commit();
-            newFeedback();
         } catch (Exception $exception) {
             DB::rollBack();
             newFeedback('پیام', 'عملیات با شکست مواجه شد', 'error');
